Move investigation selectors to the proper directory
[#22]
