+++ app/javascript/investigations/RelationshipsCardContainer.jsx
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux'
 import RelationshipsCard from 'investigations/RelationshipsCard'
-import {getPeopleSelector} from 'selectors/relationshipsSelectors'
+import {getPeopleSelector} from 'selectors/investigation/relationshipsSelectors'
 
 const mapStateToProps = (state, _ownProps) => ({
   areRelationshipsEmpty: getPeopleSelector(state).isEmpty(),
