+++ app/javascript/investigations/RelationshipsContainer.jsx
@@ -2,7 +2,7 @@ import {connect} from 'react-redux'
 import {Relationships} from 'investigations/Relationships'
 import {
   getPeopleSelector,
-} from 'selectors/relationshipsSelectors'
+} from 'selectors/investigation/relationshipsSelectors'
 
 const mapStateToProps = (state, _ownProps) => ({
   people: getPeopleSelector(state).toJS(),
