+++ app/javascript/selectors/investigation/allegationShowSelectors.js
@@ -1,6 +1,6 @@
 import {createSelector} from 'reselect'
 import {List, Map} from 'immutable'
-import {getInvestigationSelector} from 'selectors/investigationSelectors'
+import {getInvestigationSelector} from 'selectors/investigation/investigationSelectors'
 import {getAllegationTypesSelector} from 'selectors/systemCodeSelectors'
 import nameFormatter from 'utils/nameFormatter'
