+++ app/javascript/selectors/investigation/investigationSelectors.js
